requirements.txt
Flask==2.2.5
Pillow==9.5.0
requests==2.26.0